rename package platformusage to credits to have consistent namings
