+++ monitoring-service/src/main/java/com/epam/pipeline/monitor/monitoring/credits/handler/PlatformUsageCreditsRuleHandler.java
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package com.epam.pipeline.monitor.monitoring.platformusage.handler;
+package com.epam.pipeline.monitor.monitoring.credits.handler;
 
-import com.epam.pipeline.entity.platformusage.PlatformUsageCreditsUpdateEvent;
-import com.epam.pipeline.entity.platformusage.PlatformUsageCreditsUpdateRule;
-import com.epam.pipeline.entity.platformusage.PlatformUsageCreditsUpdateRuleType;
+import com.epam.pipeline.entity.credits.PlatformUsageCreditsUpdateEvent;
+import com.epam.pipeline.entity.credits.PlatformUsageCreditsUpdateRule;
+import com.epam.pipeline.entity.credits.PlatformUsageCreditsUpdateRuleType;
 
 import java.time.LocalDateTime;
 import java.util.List;
