+++ monitoring-service/src/main/java/com/epam/pipeline/monitor/rest/CloudPipelineAPIClient.java
@@ -27,9 +27,9 @@ import com.epam.pipeline.entity.cluster.NodeInstance;
 import com.epam.pipeline.entity.cluster.PodInstance;
 import com.epam.pipeline.entity.cluster.pool.NodePool;
 import com.epam.pipeline.entity.pipeline.PipelineRun;
-import com.epam.pipeline.vo.platformusage.PlatformUsageCreditsEventFilterVO;
-import com.epam.pipeline.entity.platformusage.PlatformUsageCreditsUpdateEvent;
-import com.epam.pipeline.entity.platformusage.PlatformUsageCreditsUpdateRule;
+import com.epam.pipeline.vo.credits.PlatformUsageCreditsEventFilterVO;
+import com.epam.pipeline.entity.credits.PlatformUsageCreditsUpdateEvent;
+import com.epam.pipeline.entity.credits.PlatformUsageCreditsUpdateRule;
 import com.epam.pipeline.entity.preference.Preference;
 import com.epam.pipeline.entity.user.PipelineUser;
 import com.epam.pipeline.entity.region.AbstractCloudRegion;
