+++ monitoring-service/src/test/java/com/epam/pipeline/monitor/monitoring/credits/PipelineRunUsageCreditsRuleHandlerTest.java
@@ -14,16 +14,16 @@
  * limitations under the License.
  */
 
-package com.epam.pipeline.monitor.monitoring.platformusage;
+package com.epam.pipeline.monitor.monitoring.credits;
 
 import com.epam.pipeline.entity.pipeline.PipelineRun;
 import com.epam.pipeline.entity.pipeline.TaskStatus;
-import com.epam.pipeline.monitor.monitoring.platformusage.handler.PipelineRunUsageCreditsRuleHandler;
-import com.epam.pipeline.vo.platformusage.PlatformUsageCreditsEventFilterVO;
-import com.epam.pipeline.entity.platformusage.PlatformUsageCreditsUpdateAction;
-import com.epam.pipeline.entity.platformusage.PlatformUsageCreditsUpdateEvent;
-import com.epam.pipeline.entity.platformusage.PlatformUsageCreditsUpdateRule;
-import com.epam.pipeline.entity.platformusage.PlatformUsageCreditsUpdateRuleType;
+import com.epam.pipeline.monitor.monitoring.credits.handler.PipelineRunUsageCreditsRuleHandler;
+import com.epam.pipeline.vo.credits.PlatformUsageCreditsEventFilterVO;
+import com.epam.pipeline.entity.credits.PlatformUsageCreditsUpdateAction;
+import com.epam.pipeline.entity.credits.PlatformUsageCreditsUpdateEvent;
+import com.epam.pipeline.entity.credits.PlatformUsageCreditsUpdateRule;
+import com.epam.pipeline.entity.credits.PlatformUsageCreditsUpdateRuleType;
 import com.epam.pipeline.vo.SecuredEntityVO;
 import com.epam.pipeline.entity.user.PipelineUser;
 import com.epam.pipeline.monitor.rest.CloudPipelineAPIClient;
